function_pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -10,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (array == NULL || action == NULL)
 		return;
@@ -18,6 +18,5 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 	{
 		action(array[i]);
 	}
-	
 }
